feat(s3s-sigv2): migrate PostSignatureV2
- Move PostSignatureV2 verbatim; extract now takes the borrowed
  multipart field table (values as lowercase-normalized names,
  per the multipart.fields() contract) instead of s3s's Multipart.
- s3s v2_check_post_signature passes multipart.fields().
- Remove the private sig_v2::post_signature_v2 module.
