+++ crates/s3s-sigv2/src/lib.rs
@@ -7,3 +7,6 @@
 
 mod authorization;
 pub use self::authorization::*;
+
+mod post_signature;
+pub use self::post_signature::*;
