+++ crates/s3s/src/sig_v2/mod.rs
@@ -12,8 +12,5 @@
 mod presigned_url_v2;
 pub use self::presigned_url_v2::*;
 
-mod post_signature_v2;
-pub use self::post_signature_v2::*;
-
 mod methods;
 pub use self::methods::*;
